Messenger/requirements.txt
fastapi~=0.115.12
firebase-admin~=6.8.0
python-dotenv~=1.1.0
bcrypt~=4.3.0
pydantic~=2.11.5
uvicorn~=0.34.2
flet~=0.28.3
httpx~=0.28.1